refactoring * moved all components to a separate folder
